thermal: fix OC event drivers CERT-C defect
It is reported that OC event drivers have defect of "evaluates to an
address that could be at negative offset of an array". This change
casts the type of the index of the array from int to unsigned int to
make them compliant with CERT C.

Bug 3512545
Bug 3571683

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2775201
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>

+++ drivers/thermal/tegra23x_oc_event.c
@@ -64,7 +64,8 @@ static ssize_t throt_en_show(struct device *dev, struct device_attribute *attr,
 		return -EINVAL;
 	}
 
-	return sprintf(buf, "%u\n", resp.throt_en[sensor_attr->index]);
+	return sprintf(buf, "%u\n",
+		       resp.throt_en[(unsigned int)sensor_attr->index]);
 }
 
 static ssize_t event_cnt_show(struct device *dev, struct device_attribute *attr,
@@ -95,7 +96,8 @@ static ssize_t event_cnt_show(struct device *dev, struct device_attribute *attr,
 		return -EINVAL;
 	}
 
-	return sprintf(buf, "%u\n", resp.event_cnt[sensor_attr->index]);
+	return sprintf(buf, "%u\n",
+		       resp.event_cnt[(unsigned int)sensor_attr->index]);
 }
 
 static SENSOR_DEVICE_ATTR_RO(oc1_throt_en, throt_en, TEGRA234_SOCTHERM_EDP_OC1);
